backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
python_dotenv